@@ -1,51 +1,80 @@
 {
   "name": "aquarius",
-  "version": "0.0.1",
-  "description": "",
+  "version": "2.0.0",
+  "description": "A highly configurable Discord chatbot",
   "main": "index.js",
-  "scripts": {
-    "start": "node index.js",
-    "test": "jest --setupTestFrameworkScriptFile ./index-test.js"
-  },
+  "repository": "ianmitchell/aquarius",
   "author": {
     "name": "Ian Mitchell",
     "email": "ian.mitchel1@live.com",
     "url": "http://ianmitchell.io"
   },
-  "repository": "ianmitchell/aquarius",
-  "license": "ISC",
+  "license": "MIT",
   "engines": {
-    "node": ">7.0.0"
+    "node": ">=10.0.0"
+  },
+  "scripts": {
+    "start": "cross-env DEBUG=*,-socket.io*,-engine*,-babel*,-send,-ref* node -r esm index.js",
+    "debug": "cross-env DEBUG=*,-socket.io*,-engine*,-babel*,-send,-ref* node --nolazy --inspect-brk=9229 -r esm index.js",
+    "lint": "eslint . --ext .js --ignore-path .gitignore",
+    "pretest": "yarn run lint",
+    "test": "jest --config ./jest.config.json --maxWorkers=1"
+  },
+  "esm": {
+    "mode": "auto",
+    "cjs": true
   },
   "dependencies": {
-    "blessed": "^0.1.81",
-    "blessed-contrib": "^4.6.5",
-    "chalk": "^1.1.3",
-    "debug": "^2.3.3",
-    "discord.js": "^11.2.1",
-    "dotenv": "^4.0.0",
-    "form-data": "^2.1.2",
-    "hearthstone-db": "^1.0.4",
-    "moment": "^2.16.0",
-    "mtgsdk": "^0.2.0",
-    "node-fetch": "^1.6.3",
-    "node-opus": "^0.2.4",
+    "chalk": "^2.3.0",
+    "cross-env": "^5.1.3",
+    "date-fns": "^2.0.0-alpha.26",
+    "debug": "^2.6.9",
+    "dedent-js": "^1.0.1",
+    "discord.js": "^11.4.2",
+    "downsize": "^0.0.8",
+    "esm": "^3.1.0",
+    "form-data": "^2.3.3",
+    "js-yaml": "^3.10.0",
+    "micro": "^9.1.0",
+    "micro-route": "^2.5.0",
+    "mongoist": "^2.0.0",
+    "next": "^7.0.2-canary.50",
+    "node-fetch": "^2.3.0",
+    "node-opus": "^0.3.1",
+    "node-vibrant": "^3.0.0",
+    "now-env": "^3.0.3",
     "parse-rss": "^1.0.3",
-    "pg": "^6.0.0",
-    "pg-hstore": "^2.3.2",
-    "sequelize": "^3.27.0",
-    "sequelize-cli": "^2.4.0",
-    "stack-trace": "0.0.9",
-    "umzug": "^1.11.0",
+    "pluralize": "^7.0.0",
+    "pretty-bytes": "^5.0.0",
+    "prop-types": "^15.6.2",
+    "raven": "^2.6.3",
+    "react": "^16.7.0",
+    "react-dom": "^16.7.0",
+    "rgb-hex": "^2.1.0",
+    "socket.io": "^2.2.0",
+    "socket.io-client": "^2.2.0",
+    "timber": "^3.1.3",
+    "turndown": "^5.0.1",
+    "unzalgo": "^2.0.5",
     "xmldom": "^0.1.27"
   },
-  "peerDependencies": {
-    "bufferutil": "^2.0.0"
-  },
   "devDependencies": {
-    "eslint": "^3.10.2",
-    "eslint-config-airbnb": "^14.1.0",
-    "eslint-plugin-import": "^2.2.0",
-    "jest": "^19.0.2"
+    "@babel/core": "^7.2.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
+    "babel-core": "^7.0.0-bridge.0",
+    "babel-eslint": "^10.0.1",
+    "babel-jest": "^23.6.0",
+    "babel-preset-env": "^1.6.1",
+    "eslint": "^5.9.0",
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-config-airbnb-base": "^13.1.0",
+    "eslint-plugin-import": "^2.8.0",
+    "eslint-plugin-jest": "^22.1.0",
+    "eslint-plugin-jsx-a11y": "^6.1.2",
+    "eslint-plugin-react": "^7.11.1",
+    "jest": "^23.6.0",
+    "jest-discord": "^0.1.0",
+    "jest-environment-discord": "^0.1.0",
+    "regenerator-runtime": "^0.13.1"
   }
 }
